+++ typed_sql/lib/src/codegen/build_code.dart
@@ -820,9 +820,15 @@ Iterable<Spec> buildTable(ParsedTable table, ParsedSchema schema) sync* {
               .map((f) {
                 final c = '$rowInstanceName.${f.name}';
                 if (f.isCustomType) {
-                  return '$c.asEncoded().equalsValue(${f.name}.toDatabase())';
+                  final encoded = '$c.asEncoded()';
+                  final value = '${f.name}.toDatabase()';
+                  return f.isNullable
+                      ? '$encoded.equalsUnlessNull(toExpr($value))'
+                      : '$encoded.equalsValue($value)';
                 }
-                return '$c.equalsValue(${f.name})';
+                return f.isNullable
+                    ? '$c.equalsUnlessNull(toExpr(${f.name}))'
+                    : '$c.equalsValue(${f.name})';
               })
               .join(' & ');
           return Method(
@@ -1008,7 +1014,7 @@ Iterable<Spec> buildTable(ParsedTable table, ParsedSchema schema) sync* {
                 \$ForGeneratedCode.subqueryTable(
                   _\$${ref.table.rowClass.name}._\$table
                 ).where((r) =>
-                  ${ref.fk.foreignKey.mapIndexed((i, fk) => 'r.${fk.name}.equals(${ref.fk.fields[i]})').join(' & ')}
+                  ${ref.fk.foreignKey.mapIndexed((i, fk) => _renderEquals('r.${fk.name}', fk.isNullable, ref.fk.fields[i], ref.fk.referencedFields[i].isNullable)).join(' & ')}
                 )
               '''),
           ),
@@ -1044,7 +1050,7 @@ Iterable<Spec> buildTable(ParsedTable table, ParsedSchema schema) sync* {
                 _\$${fk.referencedTable.rowClass.name}._\$table
               ).where(
                 (r) =>
-                  ${fk.fields.mapIndexed((i, f) => 'r.$f.equals(${fk.foreignKey[i].name})').join(' & ')}
+                  ${fk.fields.mapIndexed((i, f) => _renderEquals('r.$f', fk.referencedFields[i].isNullable, fk.foreignKey[i].name, fk.foreignKey[i].isNullable)).join(' & ')}
               ).$firstAsNotNull
             '''),
           );
@@ -1093,7 +1099,7 @@ Iterable<Spec> buildTable(ParsedTable table, ParsedSchema schema) sync* {
                 \$ForGeneratedCode.subqueryTable(
                   _\$${ref.table.rowClass.name}._\$table
                 ).where((r) =>
-                  ${ref.fk.foreignKey.mapIndexed((i, fk) => 'r.${fk.name}.equalsUnlessNull(${ref.fk.fields[i]}).asNotNull()').join(' & ')}
+                  ${ref.fk.foreignKey.mapIndexed((i, fk) => _renderEquals('r.${fk.name}', fk.isNullable, ref.fk.fields[i], true)).join(' & ')}
                 )
               '''),
           ),
@@ -1121,7 +1127,7 @@ Iterable<Spec> buildTable(ParsedTable table, ParsedSchema schema) sync* {
                 _\$${fk.referencedTable.rowClass.name}._\$table
               ).where(
                 (r) =>
-                  ${fk.fields.mapIndexed((i, f) => 'r.$f.equalsUnlessNull(${fk.foreignKey[i].name}).asNotNull()').join(' & ')}
+                  ${fk.fields.mapIndexed((i, f) => _renderEquals('r.$f', fk.referencedFields[i].isNullable, fk.foreignKey[i].name, true)).join(' & ')}
               ).first
             '''),
           );
@@ -1214,7 +1220,7 @@ Iterable<Spec> buildTable(ParsedTable table, ParsedSchema schema) sync* {
                   ..lambda = true
                   ..body = Code('''
                   on((a, b) =>
-                    ${ref.fk.fields.mapIndexed((i, f) => 'a.$f.equals(b.${ref.fk.foreignKey[i].name})').join(' & ')}
+                    ${ref.fk.fields.mapIndexed((i, f) => _renderEquals('a.$f', ref.fk.referencedFields[i].isNullable, 'b.${ref.fk.foreignKey[i].name}', ref.fk.foreignKey[i].isNullable)).join(' & ')}
                   )
                 '''),
               ),
@@ -1232,7 +1238,7 @@ Iterable<Spec> buildTable(ParsedTable table, ParsedSchema schema) sync* {
                   ..lambda = true
                   ..body = Code('''
                 on((a, b) =>
-                  ${fk.fields.mapIndexed((i, f) => 'b.$f.equals(a.${fk.foreignKey[i].name})').join(' & ')}
+                  ${fk.fields.mapIndexed((i, f) => _renderEquals('b.$f', fk.referencedFields[i].isNullable, 'a.${fk.foreignKey[i].name}', fk.foreignKey[i].isNullable)).join(' & ')}
                 )
               '''),
               );
@@ -1807,6 +1813,18 @@ String fieldType(ParsedField field) {
     return backingTypeName;
   }
   return '${field.typeName}Ext._exprType';
+}
+
+String _renderEquals(
+  String left,
+  bool leftNullable,
+  String right,
+  bool rightNullable,
+) {
+  if (!leftNullable && !rightNullable) {
+    return '$left.equals($right)';
+  }
+  return '$left.equalsUnlessNull($right)';
 }
 
 String renderExpression(Expression e) {