+++ typed_sql/test/typed_sql/expr/double_expr_test.dart
@@ -48,16 +48,6 @@ final _cases = [
     expr: toExpr(3.14).equals(toExpr(3.14)),
     expected: true,
   ),
-  (
-    name: '0.0.equals(null)',
-    expr: toExpr(0.0).equals(toExpr(null)),
-    expected: false,
-  ),
-  (
-    name: '3.14.equals(null)',
-    expr: toExpr(3.14).equals(toExpr(null)),
-    expected: false,
-  ),
 
   // Test for .equalsValue
   (
@@ -89,16 +79,6 @@ final _cases = [
     name: '3.14.equalsValue(3.14)',
     expr: toExpr(3.14).equalsValue(3.14),
     expected: true,
-  ),
-  (
-    name: '0.0.equalsValue(null)',
-    expr: toExpr(0.0).equalsValue(null),
-    expected: false,
-  ),
-  (
-    name: '3.14.equalsValue(null)',
-    expr: toExpr(3.14).equalsValue(null),
-    expected: false,
   ),
 
   // Test for .notEquals
@@ -132,16 +112,6 @@ final _cases = [
     expr: toExpr(3.14).notEquals(toExpr(3.14)),
     expected: false,
   ),
-  (
-    name: '0.0.notEquals(null)',
-    expr: toExpr(0.0).notEquals(toExpr(null)),
-    expected: true,
-  ),
-  (
-    name: '3.14.notEquals(null)',
-    expr: toExpr(3.14).notEquals(toExpr(null)),
-    expected: true,
-  ),
 
   // Test for .notEqualsValue
   (
@@ -173,16 +143,6 @@ final _cases = [
     name: '3.14.notEqualsValue(3.14)',
     expr: toExpr(3.14).notEqualsValue(3.14),
     expected: false,
-  ),
-  (
-    name: '0.0.notEqualsValue(null)',
-    expr: toExpr(0.0).notEqualsValue(null),
-    expected: true,
-  ),
-  (
-    name: '3.14.notEqualsValue(null)',
-    expr: toExpr(3.14).notEqualsValue(null),
-    expected: true,
   ),
 
   // Test for .lessThan