+++ typed_sql/test/typed_sql/expr/nullable_datetime_test.dart
@@ -76,22 +76,30 @@ final _cases =
       // Expr<DateTime?>.equals
       (
         name: 'null.asDateTime().equals(epoch)',
-        expr: toExpr(null).asDateTime().equals(toExpr(epoch)),
+        expr: toExpr(
+          null,
+        ).asDateTime().equalsUnlessNull(toExpr(epoch)).orElseValue(false),
         expected: false,
       ),
       (
         name: 'epoch.equals(epoch)',
-        expr: toExpr(epoch as DateTime?).equals(toExpr(epoch)),
+        expr: toExpr(
+          epoch as DateTime?,
+        ).equalsUnlessNull(toExpr(epoch)).orElseValue(false),
         expected: true,
       ),
       (
         name: 'null.asDateTime().equals(today)',
-        expr: toExpr(null).asDateTime().equals(toExpr(today)),
+        expr: toExpr(
+          null,
+        ).asDateTime().equalsUnlessNull(toExpr(today)).orElseValue(false),
         expected: false,
       ),
       (
         name: 'today.equals(today)',
-        expr: toExpr(today as DateTime?).equals(toExpr(today)),
+        expr: toExpr(
+          today as DateTime?,
+        ).equalsUnlessNull(toExpr(today)).orElseValue(false),
         expected: true,
       ),
       // Expr<DateTime?>.isNotDistinctFrom
@@ -169,64 +177,88 @@ final _cases =
       // Expr<DateTime?>.notEquals
       (
         name: 'null.asDateTime().notEquals(epoch)',
-        expr: toExpr(null).asDateTime().notEquals(toExpr(epoch)),
+        expr: toExpr(
+          null,
+        ).asDateTime().equalsUnlessNull(toExpr(epoch)).orElseValue(false).not(),
         expected: true,
       ),
       (
         name: 'epoch.notEquals(epoch)',
-        expr: toExpr(epoch as DateTime?).notEquals(toExpr(epoch)),
+        expr: toExpr(
+          epoch as DateTime?,
+        ).equalsUnlessNull(toExpr(epoch)).orElseValue(false).not(),
         expected: false,
       ),
       (
         name: 'null.asDateTime().notEquals(today)',
-        expr: toExpr(null).asDateTime().notEquals(toExpr(today)),
+        expr: toExpr(
+          null,
+        ).asDateTime().equalsUnlessNull(toExpr(today)).orElseValue(false).not(),
         expected: true,
       ),
       (
         name: 'today.notEquals(today)',
-        expr: toExpr(today as DateTime?).notEquals(toExpr(today)),
+        expr: toExpr(
+          today as DateTime?,
+        ).equalsUnlessNull(toExpr(today)).orElseValue(false).not(),
         expected: false,
       ),
       // Expr<DateTime?>.equalsValue
       (
         name: 'null.asDateTime().equalsValue(epoch)',
-        expr: toExpr(null).asDateTime().equalsValue(epoch),
+        expr: toExpr(
+          null,
+        ).asDateTime().equalsUnlessNull(toExpr(epoch)).orElseValue(false),
         expected: false,
       ),
       (
         name: 'epoch.equalsValue(epoch)',
-        expr: toExpr(epoch as DateTime?).equalsValue(epoch),
+        expr: toExpr(
+          epoch as DateTime?,
+        ).equalsUnlessNull(toExpr(epoch)).orElseValue(false),
         expected: true,
       ),
       (
         name: 'null.asDateTime().equalsValue(today)',
-        expr: toExpr(null).asDateTime().equalsValue(today),
+        expr: toExpr(
+          null,
+        ).asDateTime().equalsUnlessNull(toExpr(today)).orElseValue(false),
         expected: false,
       ),
       (
         name: 'today.equalsValue(today)',
-        expr: toExpr(today as DateTime?).equalsValue(today),
+        expr: toExpr(
+          today as DateTime?,
+        ).equalsUnlessNull(toExpr(today)).orElseValue(false),
         expected: true,
       ),
       // Expr<DateTime?>.notEqualsValue
       (
         name: 'null.asDateTime().notEqualsValue(epoch)',
-        expr: toExpr(null).asDateTime().notEqualsValue(epoch),
+        expr: toExpr(
+          null,
+        ).asDateTime().equalsUnlessNull(toExpr(epoch)).orElseValue(false).not(),
         expected: true,
       ),
       (
         name: 'epoch.notEqualsValue(epoch)',
-        expr: toExpr(epoch as DateTime?).notEqualsValue(epoch),
+        expr: toExpr(
+          epoch as DateTime?,
+        ).equalsUnlessNull(toExpr(epoch)).orElseValue(false).not(),
         expected: false,
       ),
       (
         name: 'null.asDateTime().notEqualsValue(today)',
-        expr: toExpr(null).asDateTime().notEqualsValue(today),
+        expr: toExpr(
+          null,
+        ).asDateTime().equalsUnlessNull(toExpr(today)).orElseValue(false).not(),
         expected: true,
       ),
       (
         name: 'today.notEqualsValue(today)',
-        expr: toExpr(today as DateTime?).notEqualsValue(today),
+        expr: toExpr(
+          today as DateTime?,
+        ).equalsUnlessNull(toExpr(today)).orElseValue(false).not(),
         expected: false,
       ),
       // Expr<DateTime?>.isNull()