+++ typed_sql/test/typed_sql/join/join_query/join_query_test.dart
@@ -88,7 +88,9 @@ void main() {
     final result = await db.employees
         .where((e) => e.name.endsWithValue('e'))
         .join(db.departments)
-        .on((e, d) => e.departmentId.equals(d.departmentId))
+        .on(
+          (e, d) => e.departmentId.equalsUnlessNull(d.departmentId),
+        )
         .fetch();
     check(result).length.equals(2);
   });
@@ -97,7 +99,9 @@ void main() {
     final result = await db.employees
         .where((e) => e.name.endsWithValue('e'))
         .leftJoin(db.departments)
-        .on((e, d) => e.departmentId.equals(d.departmentId))
+        .on(
+          (e, d) => e.departmentId.equalsUnlessNull(d.departmentId),
+        )
         .fetch();
     check(result).length.equals(3);
   });
@@ -106,7 +110,9 @@ void main() {
     final result = await db.employees
         .where((e) => e.name.endsWithValue('e'))
         .rightJoin(db.departments)
-        .on((e, d) => e.departmentId.equals(d.departmentId))
+        .on(
+          (e, d) => e.departmentId.equalsUnlessNull(d.departmentId),
+        )
         .fetch();
     check(result).length.equals(4);
   });
@@ -116,7 +122,9 @@ void main() {
   r.addTest('departments.join(employees.where(..))', (db) async {
     final result = await db.departments
         .join(db.employees.where((e) => e.name.endsWithValue('e')))
-        .on((d, e) => e.departmentId.equals(d.departmentId))
+        .on(
+          (d, e) => e.departmentId.equalsUnlessNull(d.departmentId),
+        )
         .fetch();
     check(result).length.equals(2);
   });
@@ -124,7 +132,9 @@ void main() {
   r.addTest('departments.leftJoin(employees.where(..))', (db) async {
     final result = await db.departments
         .leftJoin(db.employees.where((e) => e.name.endsWithValue('e')))
-        .on((d, e) => e.departmentId.equals(d.departmentId))
+        .on(
+          (d, e) => e.departmentId.equalsUnlessNull(d.departmentId),
+        )
         .fetch();
     check(result).length.equals(4);
   });
@@ -132,7 +142,9 @@ void main() {
   r.addTest('departments.rightJoin(employees.where(..))', (db) async {
     final result = await db.departments
         .rightJoin(db.employees.where((e) => e.name.endsWithValue('e')))
-        .on((d, e) => e.departmentId.equals(d.departmentId))
+        .on(
+          (d, e) => e.departmentId.equalsUnlessNull(d.departmentId),
+        )
         .fetch();
     check(result).length.equals(3);
   });
@@ -143,7 +155,9 @@ void main() {
     final result = await db.departments
         .where((d) => d.location.notEqualsValue('Floor 3'))
         .join(db.employees.where((e) => e.name.endsWithValue('e')))
-        .on((d, e) => e.departmentId.equals(d.departmentId))
+        .on(
+          (d, e) => e.departmentId.equalsUnlessNull(d.departmentId),
+        )
         .fetch();
     check(result).length.equals(2);
   });
@@ -152,7 +166,9 @@ void main() {
     final result = await db.departments
         .where((d) => d.location.notEqualsValue('Floor 3'))
         .leftJoin(db.employees.where((e) => e.name.endsWithValue('e')))
-        .on((d, e) => e.departmentId.equals(d.departmentId))
+        .on(
+          (d, e) => e.departmentId.equalsUnlessNull(d.departmentId),
+        )
         .fetch();
     check(result).length.equals(3);
   });
@@ -161,7 +177,9 @@ void main() {
     final result = await db.departments
         .where((d) => d.location.notEqualsValue('Floor 3'))
         .rightJoin(db.employees.where((e) => e.name.endsWithValue('e')))
-        .on((d, e) => e.departmentId.equals(d.departmentId))
+        .on(
+          (d, e) => e.departmentId.equalsUnlessNull(d.departmentId),
+        )
         .fetch();
     check(result).length.equals(3);
   });