+++ typed_sql/test/typed_sql/references/nullable_reference/nullable_reference_test.dart
@@ -302,7 +302,9 @@ void main() {
     (db) async {
       final result = await db.books
           .join(db.authors)
-          .on((book, author) => book.editorId.equals(author.authorId))
+          .on(
+            (book, author) => book.editorId.equalsUnlessNull(author.authorId),
+          )
           .select(
             (book, editor) => (
               book.title,